+++ src/clj/cascalog/ops.clj
@@ -215,6 +215,12 @@
     (limit [amt] :<< ?invars :>> ?outvars)
     ))
 
+(defn fixed-sample-agg [amt]
+  (<- [:<< ?invars :>> ?outvars]
+      ((cascalog.ops.RandInt.) :<< ?invars :> ?rand)
+      (:sort ?rand)
+      (limit [amt] :<< ?invars :>> ?outvars)))
+
 ;; Common patterns
 
 (defn lazy-generator