+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 
 * Fixed bug where KryoService would fail when two jobs were called simultaneously from different threads (as with cascalog.checkpoint).
 * Added `lfs-delimited` and `hfs-delimited` tap functions.
+* Added `fixed-sample` operator to `cascalog.ops`
+* Bug fix: make vars containing predmacros work properly
 
 ## 1.8.6